requirements.txt
firebase-admin>=6.0.0
overpy>=0.6
geopy>=2.3.0
PyYAML>=6.0
requests>=2.28.0